Lab6_assignment/requirements.txt
fastapi 
pydantic
dotenv
uvicorn
gunicorn
google-cloud-storage
google-cloud-bigquery